requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
rich==13.7.0
